Split SvgElementStyle
